removed few comments, seperated product Models & interfaces
