+++ HenriksHobbyLager/Interfaces/IProductFacade.cs
@@ -1,0 +1,12 @@
+﻿namespace HenriksHobbyLager
+{
+    public interface IProductFacade
+    {
+        IEnumerable<Product> GetAllProducts();
+        Product GetProduct(int id);
+        void CreateProduct(Product product);
+        void UpdateProduct(Product product);
+        void DeleteProduct(int id);
+        IEnumerable<Product> SearchProducts(string searchTerm);
+    }
+}
