+++ HenriksHobbyLager/Interfaces/IRepository.cs
@@ -1,0 +1,12 @@
+﻿namespace HenriksHobbyLager
+{
+    public interface IRepository<T>
+    {
+        IEnumerable<T> GetAll();
+        T GetById(int id);
+        void Add(T entity);
+        void Update(T entity);
+        void Delete(int id);
+        IEnumerable<T> Search(Func<T, bool> predicate);
+    }
+}
